# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface LoginRequest {
  emailAddress: string;
  /** @minLength 1 */
  password: string;
}

export interface RefreshSessionRequest {
  /** @minLength 1 */
  refreshToken: string;
}

export interface LogoutRequest {
  /** @minLength 1 */
  refreshToken: string;
}

export interface AuthTokens {
  accessToken: string;
  refreshToken: string;
}

export type LoginData = AuthTokens & {
  userId: string;
};

/**
 * Lowercase role returned by the API.
 */
export type CurrentSessionUserRole = typeof CurrentSessionUserRole[keyof typeof CurrentSessionUserRole];


export const CurrentSessionUserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface CurrentSessionUser {
  id: string;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  emailAddress: string;
  /** @nullable */
  phoneNumber: string | null;
  /** @nullable */
  avatarUrl: string | null;
  /** Lowercase role returned by the API. */
  role: CurrentSessionUserRole;
}

export interface SuccessEnvelope {
  success: true;
}

export type LoginResponse = SuccessEnvelope & {
  data: LoginData;
};

export type RefreshSessionResponse = SuccessEnvelope & {
  data: AuthTokens;
};

export type CurrentSessionResponse = SuccessEnvelope & {
  data: CurrentSessionUser;
};

export interface HealthStatus {
  status: string;
}
export interface Money {
  amountCents: number;
  currency: string;
}

export type StaffRefundProvider = typeof StaffRefundProvider[keyof typeof StaffRefundProvider];


export const StaffRefundProvider = {
  square: 'square',
} as const;

export type StaffRefundProviderStatus = typeof StaffRefundProviderStatus[keyof typeof StaffRefundProviderStatus];


export const StaffRefundProviderStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  REJECTED: 'REJECTED',
  FAILED: 'FAILED',
} as const;

export interface StaffRefund {
  id: string;
  paymentId: string;
  actorAdminId: string;
  provider: StaffRefundProvider;
  /** @nullable */
  providerRefundId: string | null;
  providerStatus: StaffRefundProviderStatus;
  reason: string;
  amount: Money;
  requestId: string;
  /** When true, the complete order inventory is restored once across all refunds or cancellation; partial per-line restoration is not supported. */
  restoreInventory: boolean;
  /** True when this refund was the durable trigger for the order's one-time complete inventory restoration. */
  inventoryRestored: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface StaffPaymentLineItem {
  productName: string;
  /** @minimum 1 */
  quantity: number;
  /** @nullable */
  selectedSides: string | null;
  options: string[];
}

export interface StaffPayment {
  id: string;
  orderId: string;
  /** @nullable */
  orderNumber: string | null;
  /** Current local kiosk order status. */
  orderStatus: string;
  provider: string;
  localStatus: string;
  /** @nullable */
  liveProviderStatus: string | null;
  amount: Money;
  refundedAmount: Money;
  refundableAmount: Money;
  /** @nullable */
  receiptUrl: string | null;
  /** @nullable */
  customerName: string | null;
  /** @nullable */
  customerPhone: string | null;
  /** Snapshot of kiosk order items and selected options. */
  items: StaffPaymentLineItem[];
  createdAt: string;
  /** @nullable */
  capturedAt: string | null;
  canCancel: boolean;
  canRefund: boolean;
  refunds: StaffRefund[];
}

export interface StaffPaymentDashboard {
  pendingCount: number;
  capturedCount: number;
  failedCount: number;
  refundedCount: number;
  capturedTotal: Money;
  refundedTotal: Money;
}

export interface CreateStaffRefundRequest {
  requestId: string;
  /**
     * @minLength 3
     * @maxLength 192
     */
  reason: string;
  /** @minimum 1 */
  amountCents: number;
  /** Restore the complete order inventory once if Square confirms this refund. This endpoint does not support per-line inventory restoration. */
  restoreInventory: boolean;
}

export type StaffPaymentResponse = SuccessEnvelope & {
  data: StaffPayment;
};

export type StaffRefundResponse = SuccessEnvelope & {
  data: StaffRefund;
};

export type StaffPaymentDashboardResponse = SuccessEnvelope & {
  data: StaffPaymentDashboard;
};

export interface PaginationMeta {
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type StaffPaymentListResponse = SuccessEnvelope & {
  data: StaffPayment[];
  meta: PaginationMeta;
};

export type ListStaffPaymentsParams = {
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
status?: ListStaffPaymentsStatus;
/**
 * @maxLength 100
 */
search?: string;
from?: string;
to?: string;
};

export type ListStaffPaymentsStatus = typeof ListStaffPaymentsStatus[keyof typeof ListStaffPaymentsStatus];


export const ListStaffPaymentsStatus = {
  pending: 'pending',
  captured: 'captured',
  failed: 'failed',
  canceled: 'canceled',
  refunded: 'refunded',
  partially_refunded: 'partially_refunded',
} as const;